test: Switch default branch to explicit 'main'
The default Git branch for a new repository can be set from the user
config (init.defaultBranch), so for predictable test results the name
must be set from the 'git init' command line. The test would fail (with
a mysterious error message about a directory not being found) if the
user config had a non-"master" default branch setting.

Following the direction of the development community, we are renaming
the default branch from 'master' to 'main'.